Remove retired retriever pipeline CLI
