+++ nemo_retriever/developer_docs/README.md
@@ -12,4 +12,3 @@ architecture, subsystems, and developer-facing tools.
 | [Retriever Harness README](../harness/README.md) | Operator and agent instructions for the artifact-first Retriever benchmark harness. |
 | [Retriever Harness PRD](harness_retriever_ingest_query_prd.md) | Product requirements for the artifact-first Retriever ingest/query benchmark harness revamp. |
 | [Root Ingest CLI Design](root_ingest_cli_design.md) | Reviewer guide for the `retriever ingest` local, batch, and service CLI ownership split. |
-| [VDB Retrieval Refactor Scope](vdb_retrieval_refactor.md) | Motivation, in-scope paths, and ownership boundaries for graph-pipeline VDB-agnostic retrieval. |
