+++ nemo_retriever/developer_docs/root_ingest_cli_design.md
@@ -215,20 +215,9 @@ image-store URI, dry-run, and quiet output.
 Service-backed query support belongs in the query CLI/service boundary, not in
 the ingest CLI.
 
-## Pipeline Compatibility
-
-`retriever pipeline run` is not the future public ingest interface. It remains
-callable, but hidden from root help, for compatibility and development behavior
-such as:
-
-- intermediate Parquet artifacts
-- pipeline reports and runtime metrics
-- eval, recall, harness, BEIR/QA workflows
-- legacy callers not yet migrated to root ingest/query
-
-For graph ingest paths, pipeline compatibility should continue to reuse the
-canonical ingest plan/execution layer instead of shelling out to root CLI
-commands.
+The supported command-line seam is `retriever ingest` for ingestion and
+`retriever query` for retrieval. Graph ingest paths reuse the canonical ingest
+plan and execution modules rather than duplicating CLI option handling.
 
 ## Adding Or Changing A Flag
