+++ nemo_retriever/src/nemo_retriever/tools/recall/beir_eval.py
@@ -5,7 +5,7 @@
 
 Reuses the existing evaluation logic from ``nemo_retriever.recall.beir`` and
 prints the standard run summary via
-``nemo_retriever.utils.detection_summary.print_run_summary``.
+``nemo_retriever.common.detection_summary.print_run_summary``.
 """
 
 from __future__ import annotations
@@ -32,8 +32,8 @@ class BEIREvaluatorActor:
     """Designer BEIR evaluation node against an existing LanceDB table.
 
     Assumes vectors were already written (for example via
-    :class:`~nemo_retriever.vdb.operators.IngestVdbOperator` or the ``retriever
-    pipeline`` upload path). After evaluation, calls ``print_run_summary`` like
+    :class:`~nemo_retriever.vdb.operators.IngestVdbOperator` or ``retriever
+    ingest``). After evaluation, calls ``print_run_summary`` like
     the batch pipeline.
     """
