+++ nemo_retriever/src/nemo_retriever/tools/recall/recall_eval.py
@@ -5,7 +5,7 @@
 
 Reuses the existing evaluation logic from ``nemo_retriever.recall.core`` and
 ``nemo_retriever.recall.beir``, and prints the standard run summary via
-``nemo_retriever.utils.detection_summary.print_run_summary``.
+``nemo_retriever.common.detection_summary.print_run_summary``.
 """
 
 from __future__ import annotations
@@ -36,8 +36,8 @@ class RecallEvaluatorActor:
     """Designer evaluation node against an existing LanceDB table.
 
     Assumes vectors were already written (for example via
-    :class:`~nemo_retriever.vdb.operators.IngestVdbOperator` or the ``retriever
-    pipeline`` upload path). Supports ``audio_recall`` (ground-truth query CSV)
+    :class:`~nemo_retriever.vdb.operators.IngestVdbOperator` or ``retriever
+    ingest``). Supports ``audio_recall`` (ground-truth query CSV)
     and ``beir`` (HuggingFace BEIR dataset) modes, then calls
     ``print_run_summary`` like the batch pipeline.
     """
