--- nemo_retriever/tests/test_graph_pipeline_cli.py
@@ -1,768 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2025, NVIDIA CORPORATION & AFFILIATES.
-# All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-import inspect
-import json
-import sys
-from types import SimpleNamespace
-from typing import Any
-
-import pandas as pd
-import pytest
-from typer.testing import CliRunner
-
-import nemo_retriever.cli.pipeline.__main__ as pipeline_main
-import nemo_retriever.common.detection_summary as detection_summary_module
-import nemo_retriever.examples.graph_pipeline as batch_pipeline
-import nemo_retriever.ingest.execution as ingest_execution
-import nemo_retriever.models as model_module
-import nemo_retriever.tools.recall.beir as beir_module
-from nemo_retriever.common.input_files import resolve_input_patterns
-
-RUNNER = CliRunner()
-
-
-def test_staged_pipeline_caption_default_and_help_are_omni() -> None:
-    option = inspect.signature(pipeline_main.run).parameters["caption_model_name"].default
-    result = RUNNER.invoke(pipeline_main.app, ["run", "--help"])
-
-    assert option.default == "nvidia/Nemotron-3-Nano-Omni-30B-A3B-Reasoning-BF16"
-    assert result.exit_code == 0
-    assert "Omni 30B BF16" in result.output
-
-
-class _FakeDataset:
-    def materialize(self):
-        return self
-
-    def take_all(self):
-        return []
-
-    def to_pandas(self):
-        return pd.DataFrame()
-
-    def groupby(self, _key):
-        class _FakeGrouped:
-            @staticmethod
-            def count():
-                class _FakeCounted:
-                    @staticmethod
-                    def count():
-                        return 1
-
-                return _FakeCounted()
-
-        return _FakeGrouped()
-
-
-class _FakeErrorRows:
-    def materialize(self):
-        return self
-
-    def count(self) -> int:
-        return 0
-
-
-class _FakeIngestor:
-    def __init__(self) -> None:
-        self.extract_params = None
-        self.extract_kwargs = {}
-        self.audio_extract_params = None
-        self.audio_asr_params = None
-        self.embed_params = None
-        self.file_patterns = None
-
-    def files(self, file_patterns):
-        self.file_patterns = file_patterns
-        return self
-
-    def extract(self, params=None, **kwargs):
-        self.extract_params = params
-        self.extract_kwargs = kwargs
-        self.audio_extract_params = kwargs.get("audio_chunk_params")
-        self.audio_asr_params = kwargs.get("asr_params")
-        return self
-
-    def extract_image_files(self, params):
-        self.extract_params = params
-        return self
-
-    def extract_audio(self, params=None, asr_params=None):
-        self.audio_extract_params = params
-        self.audio_asr_params = asr_params
-        return self
-
-    def extract_txt(self, params):
-        return self
-
-    def extract_html(self, params):
-        return self
-
-    def split(self, params):
-        return self
-
-    def embed(self, params):
-        self.embed_params = params
-        return self
-
-    def vdb_upload(self, params):
-        """In-graph VDB stage; real :class:`GraphIngestor` chains this before :meth:`ingest`."""
-        self.vdb_upload_params = params
-        return self
-
-    def ingest(self, params=None):
-        return _FakeDataset()
-
-    def get_error_rows(self, dataset=None):
-        return _FakeErrorRows()
-
-
-def _install_fake_lancedb(monkeypatch, *, row_count: int = 0) -> None:
-    class _FakeTable:
-        def count_rows(self) -> int:
-            return row_count
-
-    class _FakeDb:
-        def open_table(self, _name):
-            return _FakeTable()
-
-    monkeypatch.setitem(sys.modules, "lancedb", SimpleNamespace(connect=lambda _uri: _FakeDb()))
-
-
-def test_resolve_input_file_patterns_recurses_for_directory_inputs(tmp_path) -> None:
-    dataset_dir = tmp_path / "earnings_consulting"
-    dataset_dir.mkdir()
-
-    pdf_patterns = resolve_input_patterns(dataset_dir, "pdf")
-    txt_patterns = resolve_input_patterns(dataset_dir, "txt")
-    doc_patterns = resolve_input_patterns(dataset_dir, "doc")
-
-    assert pdf_patterns == [str(dataset_dir / "**" / "*.pdf")]
-    assert txt_patterns == [
-        str(dataset_dir / "**" / "*.txt"),
-        str(dataset_dir / "**" / "*.md"),
-        str(dataset_dir / "**" / "*.json"),
-        str(dataset_dir / "**" / "*.sh"),
-    ]
-    assert doc_patterns == [str(dataset_dir / "**" / "*.docx"), str(dataset_dir / "**" / "*.pptx")]
-
-
-def test_graph_pipeline_resolves_nested_pdf_directories(tmp_path) -> None:
-    dataset_dir = tmp_path / "earnings_consulting"
-    nested_dir = dataset_dir / "amazon_earnings_call"
-    nested_dir.mkdir(parents=True)
-    (nested_dir / "sample.pdf").write_text("placeholder", encoding="utf-8")
-
-    patterns = batch_pipeline._resolve_file_patterns(dataset_dir, "pdf")
-
-    assert patterns == [str(dataset_dir / "**" / "*.pdf")]
-
-
-def test_graph_pipeline_cli_accepts_multimodal_embed_and_page_image_flags(tmp_path, monkeypatch) -> None:
-    dataset_dir = tmp_path / "dataset"
-    dataset_dir.mkdir()
-    (dataset_dir / "sample.pdf").write_text("placeholder", encoding="utf-8")
-    missing_query_csv = tmp_path / "missing.csv"
-
-    fake_ingestor = _FakeIngestor()
-    monkeypatch.setattr(ingest_execution, "create_ingestor", lambda **_kwargs: fake_ingestor)
-    monkeypatch.setitem(
-        sys.modules,
-        "ray",
-        SimpleNamespace(shutdown=lambda: None, is_initialized=lambda: True),
-    )
-
-    _install_fake_lancedb(monkeypatch)
-    monkeypatch.setattr(model_module, "resolve_embed_model", lambda _name: "fake-embed-model")
-    result = RUNNER.invoke(
-        batch_pipeline.app,
-        [
-            str(dataset_dir),
-            "--evaluation-mode",
-            "beir",
-            "--beir-loader",
-            "vidore_hf",
-            "--beir-dataset-name",
-            "vidore_v3_computer_science",
-            "--query-csv",
-            str(missing_query_csv),
-            "--embed-modality",
-            "text_image",
-            "--embed-granularity",
-            "page",
-            "--extract-infographics",
-            "--no-extract-page-as-image",
-        ],
-    )
-
-    assert result.exit_code == 0
-    assert isinstance(fake_ingestor.file_patterns, list)
-    assert fake_ingestor.extract_params.extract_infographics is True
-    assert fake_ingestor.extract_params.extract_page_as_image is False
-    assert fake_ingestor.embed_params.embed_modality == "text_image"
-    assert fake_ingestor.embed_params.embed_granularity == "page"
-
-
-def test_graph_pipeline_cli_defaults_vdb_overwrite(tmp_path, monkeypatch) -> None:
-    dataset_dir = tmp_path / "dataset"
-    dataset_dir.mkdir()
-    (dataset_dir / "sample.pdf").write_text("placeholder", encoding="utf-8")
-
-    fake_ingestor = _FakeIngestor()
-    create_calls: list[dict[str, Any]] = []
-
-    def fake_create_ingestor(**kwargs: Any) -> _FakeIngestor:
-        create_calls.append(kwargs)
-        return fake_ingestor
-
-    monkeypatch.setattr(ingest_execution, "create_ingestor", fake_create_ingestor)
-    monkeypatch.setitem(
-        sys.modules,
-        "ray",
-        SimpleNamespace(shutdown=lambda: None, is_initialized=lambda: True),
-    )
-    monkeypatch.setattr(model_module, "resolve_embed_model", lambda _name: "fake-embed-model")
-
-    result = RUNNER.invoke(batch_pipeline.app, [str(dataset_dir), "--evaluation-mode", "none"])
-
-    assert result.exit_code == 0
-    assert create_calls == [{"run_mode": "inprocess", "ray_log_to_driver": True}]
-    assert fake_ingestor.extract_params.dpi == 200
-    assert fake_ingestor.extract_params.extract_infographics is False
-    assert fake_ingestor.extract_params.extract_page_as_image is True
-    assert fake_ingestor.vdb_upload_params.vdb_kwargs["overwrite"] is True
-
-
-def test_graph_pipeline_cli_vdb_append_forwards_overwrite_false(tmp_path, monkeypatch) -> None:
-    dataset_dir = tmp_path / "dataset"
-    dataset_dir.mkdir()
-    (dataset_dir / "sample.pdf").write_text("placeholder", encoding="utf-8")
-
-    fake_ingestor = _FakeIngestor()
-    monkeypatch.setattr(ingest_execution, "create_ingestor", lambda **_kwargs: fake_ingestor)
-    monkeypatch.setitem(
-        sys.modules,
-        "ray",
-        SimpleNamespace(shutdown=lambda: None, is_initialized=lambda: True),
-    )
-    monkeypatch.setattr(model_module, "resolve_embed_model", lambda _name: "fake-embed-model")
-
-    result = RUNNER.invoke(batch_pipeline.app, [str(dataset_dir), "--evaluation-mode", "none", "--vdb-append"])
-
-    assert result.exit_code == 0
-    assert fake_ingestor.vdb_upload_params.vdb_kwargs["overwrite"] is False
-
-
-def test_graph_pipeline_cli_vdb_flag_overrides_json(tmp_path, monkeypatch) -> None:
-    dataset_dir = tmp_path / "dataset"
-    dataset_dir.mkdir()
-    (dataset_dir / "sample.pdf").write_text("placeholder", encoding="utf-8")
-
-    fake_ingestor = _FakeIngestor()
-    monkeypatch.setattr(ingest_execution, "create_ingestor", lambda **_kwargs: fake_ingestor)
-    monkeypatch.setitem(
-        sys.modules,
-        "ray",
-        SimpleNamespace(shutdown=lambda: None, is_initialized=lambda: True),
-    )
-    monkeypatch.setattr(model_module, "resolve_embed_model", lambda _name: "fake-embed-model")
-
-    result = RUNNER.invoke(
-        batch_pipeline.app,
-        [
-            str(dataset_dir),
-            "--evaluation-mode",
-            "none",
-            "--vdb-kwargs-json",
-            json.dumps({"uri": "/tmp/custom-lancedb", "overwrite": True}),
-            "--vdb-append",
-        ],
-    )
-
-    assert result.exit_code == 0
-    assert fake_ingestor.vdb_upload_params.vdb_kwargs == {
-        "uri": "/tmp/custom-lancedb",
-        "overwrite": False,
-    }
-
-
-def test_graph_pipeline_cli_routes_audio_input_to_audio_ingestor(tmp_path, monkeypatch) -> None:
-    dataset_dir = tmp_path / "dataset"
-    dataset_dir.mkdir()
-    (dataset_dir / "sample.mp3").write_text("placeholder", encoding="utf-8")
-    missing_query_csv = tmp_path / "missing.csv"
-
-    fake_ingestor = _FakeIngestor()
-    monkeypatch.setattr(ingest_execution, "create_ingestor", lambda **_kwargs: fake_ingestor)
-    monkeypatch.setitem(
-        sys.modules,
-        "ray",
-        SimpleNamespace(shutdown=lambda: None, is_initialized=lambda: True),
-    )
-
-    _install_fake_lancedb(monkeypatch)
-    monkeypatch.setattr(model_module, "resolve_embed_model", lambda _name: "fake-embed-model")
-
-    result = RUNNER.invoke(
-        batch_pipeline.app,
-        [
-            str(dataset_dir),
-            "--input-type",
-            "audio",
-            "--evaluation-mode",
-            "audio_recall",
-            "--query-csv",
-            str(missing_query_csv),
-            "--recall-match-mode",
-            "audio_segment",
-            "--audio-match-tolerance-secs",
-            "3.0",
-            "--segment-audio",
-            "--audio-split-type",
-            "time",
-            "--audio-split-interval",
-            "45",
-        ],
-    )
-
-    assert result.exit_code == 0
-    assert isinstance(fake_ingestor.file_patterns, list)
-    assert fake_ingestor.audio_extract_params.split_type == "time"
-    assert fake_ingestor.audio_extract_params.split_interval == 45
-    assert fake_ingestor.audio_asr_params.segment_audio is True
-
-
-@pytest.mark.parametrize(
-    ("input_type", "filename", "param_key"),
-    [
-        ("txt", "sample.txt", "text_params"),
-        ("html", "sample.html", "html_params"),
-    ],
-)
-def test_graph_pipeline_cli_text_chunk_for_text_inputs_uses_dedicated_params_only(
-    tmp_path,
-    monkeypatch,
-    input_type: str,
-    filename: str,
-    param_key: str,
-) -> None:
-    dataset_dir = tmp_path / "dataset"
-    dataset_dir.mkdir()
-    (dataset_dir / filename).write_text("placeholder", encoding="utf-8")
-
-    fake_ingestor = _FakeIngestor()
-    monkeypatch.setattr(ingest_execution, "create_ingestor", lambda **_kwargs: fake_ingestor)
-    monkeypatch.setitem(
-        sys.modules,
-        "ray",
-        SimpleNamespace(shutdown=lambda: None, is_initialized=lambda: True),
-    )
-
-    _install_fake_lancedb(monkeypatch, row_count=1)
-
-    result = RUNNER.invoke(
-        batch_pipeline.app,
-        [
-            str(dataset_dir),
-            "--input-type",
-            input_type,
-            "--run-mode",
-            "batch",
-            "--text-chunk",
-            "--text-chunk-max-tokens",
-            "64",
-            "--text-chunk-overlap-tokens",
-            "8",
-            "--evaluation-mode",
-            "none",
-        ],
-    )
-
-    assert result.exit_code == 0, result.output
-    assert fake_ingestor.extract_kwargs[param_key].max_tokens == 64
-    assert fake_ingestor.extract_kwargs[param_key].overlap_tokens == 8
-    assert "split_config" not in fake_ingestor.extract_kwargs
-
-
-def test_graph_pipeline_cli_text_chunk_auto_pdf_uses_plan_split_config(tmp_path, monkeypatch) -> None:
-    dataset_dir = tmp_path / "dataset"
-    dataset_dir.mkdir()
-    (dataset_dir / "sample.pdf").write_text("placeholder", encoding="utf-8")
-
-    fake_ingestor = _FakeIngestor()
-    monkeypatch.setattr(ingest_execution, "create_ingestor", lambda **_kwargs: fake_ingestor)
-    monkeypatch.setitem(
-        sys.modules,
-        "ray",
-        SimpleNamespace(shutdown=lambda: None, is_initialized=lambda: True),
-    )
-    _install_fake_lancedb(monkeypatch, row_count=1)
-
-    result = RUNNER.invoke(
-        batch_pipeline.app,
-        [
-            str(dataset_dir),
-            "--run-mode",
-            "batch",
-            "--text-chunk",
-            "--text-chunk-max-tokens",
-            "64",
-            "--text-chunk-overlap-tokens",
-            "8",
-            "--evaluation-mode",
-            "none",
-        ],
-    )
-
-    assert result.exit_code == 0, result.output
-    assert fake_ingestor.extract_kwargs["split_config"]["pdf"]["max_tokens"] == 64
-    assert fake_ingestor.extract_kwargs["split_config"]["pdf"]["overlap_tokens"] == 8
-
-
-def test_graph_pipeline_cli_allows_default_evaluation_for_pdf_inputs(tmp_path, monkeypatch) -> None:
-    dataset_dir = tmp_path / "dataset"
-    dataset_dir.mkdir()
-    (dataset_dir / "sample.pdf").write_text("placeholder", encoding="utf-8")
-
-    fake_ingestor = _FakeIngestor()
-    monkeypatch.setattr(ingest_execution, "create_ingestor", lambda **_kwargs: fake_ingestor)
-    monkeypatch.setitem(
-        sys.modules,
-        "ray",
-        SimpleNamespace(shutdown=lambda: None, is_initialized=lambda: True),
-    )
-    monkeypatch.setattr(model_module, "resolve_embed_model", lambda _name: "fake-embed-model")
-
-    result = RUNNER.invoke(batch_pipeline.app, [str(dataset_dir), "--input-type", "pdf"])
-
-    assert result.exit_code == 0
-    assert isinstance(fake_ingestor.file_patterns, list)
-
-
-def test_graph_pipeline_cli_rejects_invalid_recall_mode(tmp_path) -> None:
-    dataset_dir = tmp_path / "dataset"
-    dataset_dir.mkdir()
-    (dataset_dir / "sample.pdf").write_text("placeholder", encoding="utf-8")
-
-    result = RUNNER.invoke(batch_pipeline.app, [str(dataset_dir), "--evaluation-mode", "recall"])
-
-    assert result.exit_code != 0
-    assert result.exception is not None
-    assert "Unsupported --evaluation-mode: 'recall'" in str(result.exception)
-
-
-def test_graph_pipeline_cli_rejects_audio_recall_for_pdf_inputs(tmp_path) -> None:
-    dataset_dir = tmp_path / "dataset"
-    dataset_dir.mkdir()
-    (dataset_dir / "sample.pdf").write_text("placeholder", encoding="utf-8")
-
-    result = RUNNER.invoke(
-        batch_pipeline.app,
-        [str(dataset_dir), "--input-type", "pdf", "--evaluation-mode", "audio_recall"],
-    )
-
-    assert result.exit_code != 0
-    assert result.exception is not None
-    assert "--evaluation-mode=audio_recall is only supported with --input-type=audio" in str(result.exception)
-
-
-def test_graph_pipeline_cli_routes_beir_mode_to_evaluator(tmp_path, monkeypatch) -> None:
-    dataset_dir = tmp_path / "dataset"
-    dataset_dir.mkdir()
-    (dataset_dir / "sample.pdf").write_text("placeholder", encoding="utf-8")
-
-    fake_ingestor = _FakeIngestor()
-    monkeypatch.setattr(ingest_execution, "create_ingestor", lambda **_kwargs: fake_ingestor)
-    monkeypatch.setattr(pipeline_main, "_count_uploadable_vdb_records", lambda _records: 1)
-    monkeypatch.setattr(detection_summary_module, "print_run_summary", lambda *args, **kwargs: None)
-
-    _install_fake_lancedb(monkeypatch, row_count=1)
-    monkeypatch.setitem(
-        sys.modules,
-        "ray",
-        SimpleNamespace(shutdown=lambda: None, is_initialized=lambda: True),
-    )
-    monkeypatch.setattr(model_module, "resolve_embed_model", lambda _name: "fake-embed-model")
-
-    captured = {}
-
-    def _fake_evaluate(cfg):
-        captured["cfg"] = cfg
-        return type("Dataset", (), {"query_ids": ["1", "2"]})(), [], {}, {"ndcg@10": 0.75, "recall@5": 0.6}
-
-    monkeypatch.setattr(beir_module, "evaluate_lancedb_beir", _fake_evaluate)
-
-    result = RUNNER.invoke(
-        batch_pipeline.app,
-        [
-            str(dataset_dir),
-            "--evaluation-mode",
-            "beir",
-            "--beir-loader",
-            "vidore_hf",
-            "--beir-dataset-name",
-            "vidore_v3_computer_science",
-            "--beir-k",
-            "5",
-            "--beir-k",
-            "10",
-        ],
-    )
-
-    assert result.exit_code == 0
-    assert captured["cfg"].loader == "vidore_hf"
-    assert captured["cfg"].dataset_name == "vidore_v3_computer_science"
-    assert tuple(captured["cfg"].ks) == (5, 10)
-
-
-def test_graph_pipeline_cli_accepts_harness_runtime_metric_flags(tmp_path, monkeypatch) -> None:
-    dataset_dir = tmp_path / "dataset"
-    dataset_dir.mkdir()
-    (dataset_dir / "sample.pdf").write_text("placeholder", encoding="utf-8")
-    missing_query_csv = tmp_path / "missing.csv"
-    runtime_dir = tmp_path / "runtime_metrics"
-
-    fake_ingestor = _FakeIngestor()
-    monkeypatch.setattr(ingest_execution, "create_ingestor", lambda **_kwargs: fake_ingestor)
-    monkeypatch.setitem(
-        sys.modules,
-        "ray",
-        SimpleNamespace(shutdown=lambda: None, is_initialized=lambda: True),
-    )
-
-    _install_fake_lancedb(monkeypatch)
-    monkeypatch.setattr(model_module, "resolve_embed_model", lambda _name: "fake-embed-model")
-    monkeypatch.setattr(pipeline_main, "_count_uploadable_vdb_records", lambda _records: 1)
-    monkeypatch.setattr(
-        pipeline_main,
-        "_run_evaluation",
-        lambda **_kwargs: ("BEIR", 0.0, {"recall@5": 0.6}, 2, True),
-    )
-
-    result = RUNNER.invoke(
-        batch_pipeline.app,
-        [
-            str(dataset_dir),
-            "--evaluation-mode",
-            "beir",
-            "--beir-loader",
-            "vidore_hf",
-            "--beir-dataset-name",
-            "vidore_v3_computer_science",
-            "--query-csv",
-            str(missing_query_csv),
-            "--runtime-metrics-dir",
-            str(runtime_dir),
-            "--runtime-metrics-prefix",
-            "sample-run",
-            "--no-recall-details",
-        ],
-    )
-
-    assert result.exit_code == 0
-    summary_path = runtime_dir / "sample-run.runtime.summary.json"
-    assert summary_path.exists()
-    payload = json.loads(summary_path.read_text(encoding="utf-8"))
-    assert payload["recall_details"] is False
-    assert payload["evaluation_mode"] == "beir"
-
-
-def test_graph_pipeline_cli_service_mode_rejects_ingest_flag(tmp_path) -> None:
-    dataset_dir = tmp_path / "dataset"
-    dataset_dir.mkdir()
-    (dataset_dir / "sample.pdf").write_text("placeholder", encoding="utf-8")
-
-    result = RUNNER.invoke(
-        batch_pipeline.app,
-        [
-            str(dataset_dir),
-            "--run-mode",
-            "service",
-            "--ocr-invoke-url",
-            "http://localhost:9000/v1/infer",
-        ],
-    )
-
-    assert result.exit_code != 0
-    assert "--run-mode=service" in result.output
-    assert "--ocr-invoke-url" in result.output
-
-
-def test_graph_pipeline_cli_service_mode_lists_all_incompatible_flags(tmp_path) -> None:
-    dataset_dir = tmp_path / "dataset"
-    dataset_dir.mkdir()
-    (dataset_dir / "sample.pdf").write_text("placeholder", encoding="utf-8")
-
-    result = RUNNER.invoke(
-        batch_pipeline.app,
-        [
-            str(dataset_dir),
-            "--run-mode",
-            "service",
-            "--ocr-invoke-url",
-            "http://localhost:9000/v1/infer",
-            "--ray-address",
-            "ray://localhost:10001",
-            "--caption-device",
-            "cuda:0",
-        ],
-    )
-
-    assert result.exit_code != 0
-    assert "--ocr-invoke-url" in result.output
-    assert "--ray-address" in result.output
-    assert "--caption-device" in result.output
-
-
-def test_graph_pipeline_cli_service_mode_allows_extract_and_embed_flags(tmp_path, monkeypatch) -> None:
-    """Flags whose values flow through to ``ServiceIngestor`` must not be rejected."""
-    import nemo_retriever.service.service_ingestor as service_ingestor_module
-
-    dataset_dir = tmp_path / "dataset"
-    dataset_dir.mkdir()
-    (dataset_dir / "sample.pdf").write_text("placeholder", encoding="utf-8")
-    save_dir = tmp_path / "save"
-
-    captured: dict[str, Any] = {}
-
-    class _FakeServiceIngestor(list):
-        def __init__(self, *args, **kwargs) -> None:
-            super().__init__()
-
-        def files(self, _files):
-            return self
-
-        def extract(self, params=None, *, split_config=None, extraction_mode="auto", **_kwargs):
-            captured["extract_params"] = params
-            captured["split_config"] = split_config
-            captured["extraction_mode"] = extraction_mode
-            return self
-
-        def dedup(self, params=None, **_kwargs):
-            captured["dedup_params"] = params
-            return self
-
-        def embed(self, params=None, **_kwargs):
-            captured["embed_params"] = params
-            return self
-
-        def ingest(self, *args, **kwargs):
-            return self
-
-    monkeypatch.setattr(service_ingestor_module, "ServiceIngestor", _FakeServiceIngestor)
-    monkeypatch.setattr(model_module, "resolve_embed_model", lambda _name: "fake-embed-model")
-
-    result = RUNNER.invoke(
-        batch_pipeline.app,
-        [
-            str(dataset_dir),
-            "--run-mode",
-            "service",
-            "--service-url",
-            "http://localhost:7670",
-            "--embed-model-name",
-            "nvidia/llama-3.2-nv-embedqa-1b-v2",
-            "--method",
-            "ocr",
-            "--dpi",
-            "300",
-            "--no-extract-text",
-            "--embed-granularity",
-            "page",
-            "--dedup",
-            "--dedup-iou-threshold",
-            "0.6",
-            "--text-chunk",
-            "--text-chunk-max-tokens",
-            "64",
-            "--evaluation-mode",
-            "none",
-            "--save-intermediate",
-            str(save_dir),
-        ],
-    )
-
-    assert result.exit_code == 0, result.output
-    assert captured["extract_params"].method == "ocr"
-    assert captured["extract_params"].dpi == 300
-    assert captured["extract_params"].extract_text is False
-    assert captured["embed_params"].embed_granularity == "page"
-    assert captured["dedup_params"].iou_threshold == 0.6
-    assert captured["split_config"]["pdf"]["max_tokens"] == 64
-
-
-def test_graph_pipeline_cli_service_mode_rejects_vdb_flags(tmp_path) -> None:
-    dataset_dir = tmp_path / "dataset"
-    dataset_dir.mkdir()
-    (dataset_dir / "sample.pdf").write_text("placeholder", encoding="utf-8")
-
-    result_no_vdb = RUNNER.invoke(
-        batch_pipeline.app,
-        [str(dataset_dir), "--run-mode", "service", "--no-vdb"],
-    )
-    assert result_no_vdb.exit_code != 0
-    assert "--no-vdb" in result_no_vdb.output
-
-    result_overwrite = RUNNER.invoke(
-        batch_pipeline.app,
-        [str(dataset_dir), "--run-mode", "service", "--vdb-overwrite"],
-    )
-    assert result_overwrite.exit_code != 0
-    assert "--vdb-overwrite" in result_overwrite.output
-
-    result_append = RUNNER.invoke(
-        batch_pipeline.app,
-        [str(dataset_dir), "--run-mode", "service", "--vdb-append"],
-    )
-    assert result_append.exit_code != 0
-    assert "--vdb-overwrite" in result_append.output
-
-
-def test_graph_pipeline_cli_service_mode_accepts_allowlisted_flags(tmp_path, monkeypatch) -> None:
-    import nemo_retriever.service.service_ingestor as service_ingestor_module
-
-    dataset_dir = tmp_path / "dataset"
-    dataset_dir.mkdir()
-    (dataset_dir / "sample.pdf").write_text("placeholder", encoding="utf-8")
-    save_dir = tmp_path / "save"
-
-    class _FakeServiceIngestor(list):
-        def __init__(self, *args, **kwargs) -> None:
-            super().__init__()
-
-        def files(self, _files):
-            return self
-
-        def extract(self, *args, **kwargs):
-            return self
-
-        def embed(self, *args, **kwargs):
-            return self
-
-        def ingest(self, *args, **kwargs):
-            return self
-
-    monkeypatch.setattr(service_ingestor_module, "ServiceIngestor", _FakeServiceIngestor)
-    monkeypatch.setattr(model_module, "resolve_embed_model", lambda _name: "fake-embed-model")
-
-    result = RUNNER.invoke(
-        batch_pipeline.app,
-        [
-            str(dataset_dir),
-            "--run-mode",
-            "service",
-            "--service-url",
-            "http://localhost:7670",
-            "--service-concurrency",
-            "2",
-            "--embed-model-name",
-            "nvidia/llama-3.2-nv-embedqa-1b-v2",
-            "--evaluation-mode",
-            "none",
-            "--save-intermediate",
-            str(save_dir),
-        ],
-    )
-
-    assert result.exit_code == 0, result.output
