--- nemo_retriever/tests/test_pipeline_helpers.py
@@ -1,304 +1,0 @@
-# SPDX-FileCopyrightText: Copyright (c) 2024-25, NVIDIA CORPORATION & AFFILIATES.
-# All rights reserved.
-# SPDX-License-Identifier: Apache-2.0
-
-from __future__ import annotations
-
-from pathlib import Path
-from types import SimpleNamespace
-
-import pandas as pd
-import pytest
-import typer
-
-import nemo_retriever.cli.pipeline as pipeline_pkg
-from nemo_retriever.ingest.service import (
-    execute_service_ingest_request,
-    ServiceIngestRequest,
-    build_service_ingestor,
-)
-from nemo_retriever.common.params import EmbedParams, ExtractParams, TextChunkParams
-from nemo_retriever.cli.pipeline.__main__ import (
-    _build_embed_params,
-    _collect_results,
-    _count_input_units,
-    _count_uploadable_vdb_records,
-    _parse_vdb_kwargs_json,
-    _resolve_file_patterns,
-)
-from nemo_retriever.service.config import PipelineOverridesConfig
-from nemo_retriever.common.schemas.pipeline_spec import PipelineSpec
-from nemo_retriever.common.policy import validate_pipeline_spec
-from nemo_retriever.service.service_ingestor import ServiceIngestor
-
-
-def test_pipeline_package_exports_cli_app_and_run() -> None:
-    from nemo_retriever.cli.pipeline.__main__ import app, run
-
-    assert pipeline_pkg.app is app
-    assert pipeline_pkg.run is run
-    assert set(pipeline_pkg.__all__) == {"app", "run"}
-
-
-@pytest.mark.parametrize(
-    ("input_type", "files", "expected_globs"),
-    [
-        ("pdf", ["nested/doc.pdf"], ["*.pdf"]),
-        ("doc", ["deck.pptx", "report.docx"], ["*.docx", "*.pptx"]),
-        ("image", ["plot.png"], ["*.png"]),
-        ("video", ["clip.mp4"], ["*.mp4"]),
-    ],
-)
-def test_resolve_file_patterns_recurses_directory_inputs(
-    tmp_path: Path,
-    input_type: str,
-    files: list[str],
-    expected_globs: list[str],
-) -> None:
-    for name in files:
-        path = tmp_path / name
-        path.parent.mkdir(parents=True, exist_ok=True)
-        path.write_bytes(b"x")
-
-    assert _resolve_file_patterns(tmp_path, input_type) == [str(tmp_path / "**" / glob) for glob in expected_globs]
-
-
-def test_build_service_ingestor_wires_extract_embed_and_chunking(tmp_path: Path) -> None:
-    pdf = tmp_path / "doc.pdf"
-    pdf.write_bytes(b"%PDF-1.4")
-
-    ingestor = build_service_ingestor(
-        ServiceIngestRequest(
-            documents=[str(pdf)],
-            input_type="pdf",
-            extract_params=ExtractParams(method="ocr", extract_text=False, dpi=300),
-            embed_params=EmbedParams(embed_granularity="page"),
-            text_chunk_params=TextChunkParams(max_tokens=64, overlap_tokens=8),
-            enable_text_chunk=True,
-        )
-    )
-
-    assert isinstance(ingestor, ServiceIngestor)
-    payload = ingestor._pipeline_payload()
-    assert payload is not None
-    assert payload["extraction_mode"] == "pdf"
-    assert payload["extract_params"]["method"] == "ocr"
-    assert payload["extract_params"]["extract_text"] is False
-    assert payload["extract_params"]["dpi"] == 300
-    assert "batch_tuning" not in payload["extract_params"]
-    assert payload["split_config"]["pdf"]["max_tokens"] == 64
-    assert payload["split_config"]["pdf"]["overlap_tokens"] == 8
-    assert payload["embed_params"]["embed_granularity"] == "page"
-    assert "model_name" not in payload["embed_params"]
-
-    validate_pipeline_spec(
-        PipelineSpec.model_validate(ingestor._pipeline_spec),
-        PipelineOverridesConfig().to_policy(),
-    )
-
-
-def test_build_service_ingestor_does_not_forward_environment_api_key(monkeypatch, tmp_path: Path) -> None:
-    pdf = tmp_path / "doc.pdf"
-    pdf.write_bytes(b"%PDF-1.4")
-    monkeypatch.setenv("NGC_API_KEY", "environment-secret")
-    stage_params: dict[str, object] = {}
-    original_extract = ServiceIngestor.extract
-    original_embed = ServiceIngestor.embed
-
-    def capture_extract(self, params=None, **kwargs):
-        stage_params["extract"] = params
-        return original_extract(self, params, **kwargs)
-
-    def capture_embed(self, params=None, **kwargs):
-        stage_params["embed"] = params
-        return original_embed(self, params, **kwargs)
-
-    monkeypatch.setattr(ServiceIngestor, "extract", capture_extract)
-    monkeypatch.setattr(ServiceIngestor, "embed", capture_embed)
-
-    ingestor = build_service_ingestor(
-        ServiceIngestRequest(
-            documents=[str(pdf)],
-            input_type="pdf",
-            extract_params=ExtractParams(method="pdfium", use_table_structure=True),
-            embed_params=EmbedParams(embed_granularity="page"),
-        )
-    )
-
-    payload = ingestor._pipeline_payload()
-    assert payload is not None
-    assert isinstance(stage_params["extract"], ExtractParams)
-    assert isinstance(stage_params["embed"], EmbedParams)
-    assert payload["extract_params"]["method"] == "pdfium"
-    assert payload["extract_params"]["use_table_structure"] is True
-    assert payload["embed_params"]["embed_granularity"] == "page"
-    assert not ({"api_key", "page_elements_api_key", "ocr_api_key"} & payload["extract_params"].keys())
-    assert "api_key" not in payload["embed_params"]
-
-
-def test_execute_service_ingest_request_raises_for_document_failures(monkeypatch, tmp_path: Path) -> None:
-    request = ServiceIngestRequest(documents=[str(tmp_path / "doc.pdf")], input_type="pdf")
-    failed_result = SimpleNamespace(failures=[("doc.pdf", "HTTP 400: invalid request")])
-    monkeypatch.setattr(
-        "nemo_retriever.ingest.service.build_service_ingestor",
-        lambda _request: SimpleNamespace(ingest=lambda: failed_result),
-    )
-
-    with pytest.raises(RuntimeError, match=r"failed for 1 document\(s\).+doc.pdf.+HTTP 400"):
-        execute_service_ingest_request(request)
-
-
-def test_resolve_file_patterns_returns_existing_file_verbatim(tmp_path: Path) -> None:
-    path = tmp_path / "doc.pdf"
-    path.write_bytes(b"x")
-
-    assert _resolve_file_patterns(path, "audio") == [str(path)]
-
-
-def test_resolve_file_patterns_rejects_missing_or_empty_inputs(tmp_path: Path) -> None:
-    with pytest.raises(typer.BadParameter, match="Path does not exist"):
-        _resolve_file_patterns(tmp_path / "missing", "pdf")
-
-    (tmp_path / "sidecar.json").write_text("{}", encoding="utf-8")
-    with pytest.raises(typer.BadParameter, match="No files found"):
-        _resolve_file_patterns(tmp_path, "pdf")
-
-
-def test_parse_vdb_kwargs_json_keeps_backend_kwargs_opaque() -> None:
-    assert _parse_vdb_kwargs_json(None) == {}
-    assert _parse_vdb_kwargs_json('{"collection_name": "docs", "uri": "http://localhost:19530"}') == {
-        "collection_name": "docs",
-        "uri": "http://localhost:19530",
-    }
-
-
-def test_parse_vdb_kwargs_json_rejects_non_object_json() -> None:
-    with pytest.raises(typer.BadParameter, match="JSON object"):
-        _parse_vdb_kwargs_json('["not", "an", "object"]')
-
-
-def test_build_embed_params_forwards_remote_and_modality_flags() -> None:
-    params = _build_embed_params(
-        embed_model_name="nvidia/test-embed",
-        embed_invoke_url="http://embed.example/v1",
-        embed_remote_api_key="nvapi-secret",
-        embed_modality="text_image",
-        text_elements_modality="text",
-        structured_elements_modality="image",
-        embed_granularity="element",
-        embed_actors=2,
-        embed_batch_size=16,
-        embed_cpus_per_actor=1.5,
-        embed_gpus_per_actor=0.5,
-    )
-
-    assert isinstance(params, EmbedParams)
-    assert params.model_name == "nvidia/test-embed"
-    assert params.embed_invoke_url == "http://embed.example/v1"
-    assert params.api_key == "nvapi-secret"
-    assert params.embed_modality == "text_image"
-    assert params.text_elements_modality == "text"
-    assert params.structured_elements_modality == "image"
-    assert params.embed_granularity == "element"
-    assert params.inference_batch_size == 16
-    assert params.batch_tuning.embed_workers == 2
-    assert params.batch_tuning.embed_batch_size == 16
-    assert params.batch_tuning.embed_cpus_per_actor == 1.5
-    assert params.batch_tuning.gpu_embed == 0.0
-
-
-class TestCollectResults:
-    """Ingest returns a DataFrame (``ingestor.ingest()`` → ``ds.to_pandas()``); _collect_results consumes it."""
-
-    def test_batch_mode_accepts_ingest_dataframe(self):
-        rows = [
-            {"source_id": "a", "text": "hello"},
-            {"source_id": "a", "text": "world"},
-            {"source_id": "b", "text": "!"},
-        ]
-        # Same shape as the graph executor return after ``Dataset.to_pandas()``.
-        result_df = pd.DataFrame(rows)
-
-        records, df, download_time, num_units = _collect_results("batch", result_df)
-
-        assert records == rows
-        assert df is result_df
-        assert isinstance(df, pd.DataFrame)
-        assert list(df.columns) == ["source_id", "text"]
-        assert len(df) == 3
-        # ``source_id`` has two distinct values → that is the unit count.
-        assert num_units == 2
-        assert download_time >= 0.0
-
-    def test_batch_mode_handles_empty_result(self):
-        result_df = pd.DataFrame()
-        records, df, download_time, num_units = _collect_results("batch", result_df)
-        assert records == []
-        assert df.empty
-        # Empty DataFrame has no columns → falls through to len(df.index) == 0.
-        assert num_units == 0
-        assert download_time >= 0.0
-
-    def test_inprocess_mode_accepts_dataframe_directly(self):
-        rows = [
-            {"source_id": "a", "text": "x"},
-            {"source_id": "b", "text": "y"},
-        ]
-        df_in = pd.DataFrame(rows)
-
-        records, df_out, download_time, num_units = _collect_results("inprocess", df_in)
-
-        # The DataFrame is passed through unchanged (same object).
-        assert df_out is df_in
-        assert records == rows
-        # inprocess mode never incurs Ray download time.
-        assert download_time == 0.0
-        assert num_units == 2
-
-
-def test_collect_results_accepts_inprocess_dataframe() -> None:
-    df_in = pd.DataFrame([{"source_path": "/a.pdf"}, {"source_path": "/b.pdf"}])
-
-    records, df_out, download_time, num_units = _collect_results("inprocess", df_in)
-
-    assert df_out is df_in
-    assert records == [{"source_path": "/a.pdf"}, {"source_path": "/b.pdf"}]
-    assert download_time == 0.0
-    assert num_units == 2
-
-
-def test_to_client_vdb_records_returns_empty_list_when_nothing_uploadable() -> None:
-    from nemo_retriever.common.vdb.records import to_client_vdb_records
-
-    assert to_client_vdb_records([]) == []
-    assert to_client_vdb_records([{"text": "no embedding"}]) == []
-
-
-def test_count_uploadable_vdb_records_filters_rows_without_embedding_or_text() -> None:
-    rows = [
-        {
-            "text": "keep",
-            "text_embeddings_1b_v2": {"embedding": [0.1, 0.2]},
-            "source_id": "/tmp/doc-a.pdf",
-            "page_number": 1,
-        },
-        {
-            "text": "drop missing embedding",
-            "source_id": "/tmp/doc-a.pdf",
-            "page_number": 2,
-        },
-        {
-            "text_embeddings_1b_v2": {"embedding": [0.3, 0.4]},
-            "source_id": "/tmp/doc-a.pdf",
-            "page_number": 3,
-        },
-    ]
-
-    assert _count_uploadable_vdb_records(rows) == 1
-    assert _count_uploadable_vdb_records([]) == 0
-
-
-def test_count_input_units_prefers_source_id_then_source_path() -> None:
-    assert _count_input_units(pd.DataFrame({"source_id": ["a", "a", "b"]})) == 2
-    assert _count_input_units(pd.DataFrame({"source_path": ["/a", "/b", "/b"]})) == 2
-    assert _count_input_units(pd.DataFrame({"text": ["x", "y", "z"]})) == 3
